Move dotenv to dependencies to fix app startup error
We saw Google App Engine error out when attempting to start the app with
`node -r dotenv/config server/index`:

  node:internal/modules/cjs/loader:1029
    throw err;
    ^
  Error: Cannot find module 'dotenv/config'
  Require stack:
  - internal/preload
      at Function.Module._resolveFilename (node:internal/modules/cjs/loader:1026:15)
      at Function.Module._load (node:internal/modules/cjs/loader:871:27)
      at internalRequire (node:internal/modules/cjs/loader:169:19)
      at Module._preloadModules (node:internal/modules/cjs/loader:1373:5)
      at loadPreloadModules (node:internal/bootstrap/pre_execution:583:5)
      at prepareMainThreadExecution (node:internal/bootstrap/pre_execution:95:3)
      at node:internal/main/run_main_module:9:1 {
    code: 'MODULE_NOT_FOUND',
    requireStack: [ 'internal/preload' ]
  }

bin/update_gae_pkg was manually adding dotEnv to the list of
dependencies, but that seemed to have stopped working. It's unclear why;
maybe something about the GAE environment changed.

Anyway, we solved the error by moving dotEnv from devDependencies to
dependencies in package.json, and removing the code in
bin/update_gae_pkg that shuffles dotEnv from one to the other. I think
the only reason dotEnv was not in dependencies in the first place is
that it may not be necessary for Heroku deployments.